+++ src/lib/pages/contract-details/index.tsx
@@ -5,14 +5,18 @@ import {
   Tabs,
   TabPanels,
   TabPanel,
+  Icon,
+  Text,
 } from "@chakra-ui/react";
 import { observer } from "mobx-react-lite";
 import { useRouter } from "next/router";
+import { MdSearchOff } from "react-icons/md";
 
 import { BackButton } from "lib/components/button/BackButton";
 import { CustomTab } from "lib/components/CustomTab";
 import PageContainer from "lib/components/PageContainer";
-import { useContractDetail } from "lib/model/contract";
+import { useValidateAddress } from "lib/hooks";
+import { useContractData } from "lib/model/contract";
 import type { ContractAddr } from "lib/types";
 import { getFirstQueryParam, jsonPrettify } from "lib/utils";
 
@@ -23,48 +27,61 @@ import { InstantiateInfo } from "./components/InstantiateInfo";
 import { JsonInfo } from "./components/JsonInfo";
 import { TokenSection } from "./components/TokenSection";
 
-const ContractDetails = observer(() => {
-  /**
-   * @todos add contract address validation function here
-   */
-  const router = useRouter();
-  const contractAddressParam = getFirstQueryParam(
-    router.query.contractAddress
-  ) as ContractAddr;
+interface ContractDetailsBodyProps {
+  contractAddress: ContractAddr;
+}
 
-  const contractDetails = useContractDetail(contractAddressParam);
+const InvalidContract = () => (
+  <Flex
+    direction="column"
+    alignItems="center"
+    borderY="1px solid"
+    borderColor="divider.main"
+    width="full"
+    my="24px"
+    py="24px"
+  >
+    <Icon as={MdSearchOff} color="gray.600" boxSize="128" />
+    <Heading as="h5" variant="h5" my="8px">
+      Contract does not exist
+    </Heading>
+    <Text variant="body2" fontWeight="500" color="gray.500" textAlign="center">
+      Please double-check your spelling and make sure you have selected the
+      correct network.
+    </Text>
+  </Flex>
+);
 
-  // TODO - Wait for design
-  if (!contractDetails) return null;
-
+const ContractDetailsBody = ({ contractAddress }: ContractDetailsBodyProps) => {
+  const contractData = useContractData(contractAddress);
+  if (!contractData) return <InvalidContract />;
   return (
-    <PageContainer>
-      <BackButton />
-      <ContractTop contractDetail={contractDetails} />
+    <>
+      <ContractTop contractData={contractData} />
       {/* Tokens Section */}
       <TokenSection />
       {/* Contract Description Section */}
-      <ContractDesc contractDetail={contractDetails} />
+      <ContractDesc contractData={contractData} />
       {/* Query/Execute commands section */}
       <CommandSection />
       {/* Instantiate/Contract Info Section */}
       <Flex my={12} justify="space-between">
         {/* Instantiate Info */}
-        <InstantiateInfo contractDetail={contractDetails} />
+        <InstantiateInfo contractData={contractData} />
         {/* Contract Info (Expand) */}
         <Flex direction="column" flex={0.8} gap={4}>
           <JsonInfo
             header="Contract Info"
             jsonString={jsonPrettify(
               JSON.stringify(
-                contractDetails?.instantiateInfo?.raw.contract_info ?? {}
+                contractData.instantiateInfo?.raw.contract_info ?? {}
               )
             )}
             jsonAreaHeight="180px"
           />
           <JsonInfo
             header="Instantiate Messages"
-            jsonString={jsonPrettify(contractDetails?.initMsg ?? "")}
+            jsonString={jsonPrettify(contractData.initMsg ?? "")}
             showViewFullButton
             defaultExpand
           />
@@ -105,6 +122,26 @@ const ContractDetails = observer(() => {
           </TabPanel>
         </TabPanels>
       </Tabs>
+    </>
+  );
+};
+
+const ContractDetails = observer(() => {
+  const router = useRouter();
+  const { validateContractAddress } = useValidateAddress();
+
+  const contractAddressParam = getFirstQueryParam(router.query.contractAddress);
+
+  return (
+    <PageContainer>
+      <BackButton />
+      {validateContractAddress(contractAddressParam) ? (
+        <InvalidContract />
+      ) : (
+        <ContractDetailsBody
+          contractAddress={contractAddressParam as ContractAddr}
+        />
+      )}
     </PageContainer>
   );
 });